backend/requirements.txt
django>=3.2
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6
django-cors-headers>=3.6
django-environ>=0.4.5
requests>=2.22.0
python-decouple>=3.4
celery>=5.2.3
redis>=3.5.3
django-filter>=2.4
drf-writable-nested>=0.6.3
mailchimp-marketing
gunicorn>=20.0.4
mailchimp-transactional>=1.0.30
pillow
sentry-sdk>=1.3.1